backend/requirements.txt
fastapi
uvicorn[standard]
httpx
openpyxl
pillow
pydantic-settings
